+++ DatabaseApi/Models/MappableModel.cs
@@ -1,14 +1,15 @@
 // System
 using System;
+using System.Numerics;
 
 
 namespace DatabaseApi.Models
 {
     public class MappableModel
     {
-        public int Id { get; set; }
+        public BigInteger Id { get; set; }
         public string Name { get; set; }
-        public int? WebId { get; set; }
+        public BigInteger? WebId { get; set; }
         public string WebName { get; set; }
         public int? TargetWeb { get; set; }
         public DateTime? LastUpdate { get; set; }
